Make functions async so overrides can await
Swift will soon start warning about unsafe access to mutable state. These warnings can be turned on in Xcode 14 using -Xfrontend -warn-concurrency. An example of a violation is any static var on a class, for example a "shared" singleton instance.

It may not be uncommon for subclasses of APIBase to use a pattern like this to store an access token, or need to fetch some shared state from another object. The acceptable way to guard access to this shared mutable state is by marking the property with an actor attribute (such as @MainActor or a custom global actor).

Making these functions async gives them the flexibility to await these sorts of mutable shared state or call other async functions.