Add sendRequest(_:)
The old code didn't allow for an override to add code after the URLRequest was built, but before it hit the network. Adding this second flavor of sendRequest resolves this. Subclasses can now override the parameter-less sendRequest(), call buildURLRequest() within that, and then call the superclass implementation of sendRequest(_:), passing the existing URLRequest instance.

One reason for this is to perform some logging before sending the request.